Change default download directory to 'Reddit Backgrounds'
By default, Mac OS X won't display dot files in finder. This means that
~/.background-images wouldn't be selectable from System
Preferences/Desktop/Backgrounds as a folder.

The solution here is to remove the leading dot, and use a more user-friendly
name while we're at it.

The `--download-directory` CLI option and `download_directory` config allows
the user to override this to whatever they want.