Use image information provided by Reddit API
The reddit API provides image metadata directly in the API, so we can use that
instead of having to scrape that from the post title (as is common in the
SFWPorn sites, like CarPorn, MapPorn etc).

This means that we can now support non-SFWPorn reddits like /r/aww.

Thanks to @joaquimds for pointing this out and providing sample code.

Fixes #13